helloworld/requirements.txt
Flask  >= 1.1.2, < 2.0.0
gunicorn  >= 20.0.4, < 21.0.0
google-cloud-logging  >= 1.15.0, < 2.0.0
